requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
google-api-python-client==2.108.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.3.1